+++ CHANGELOG.md
@@ -1,6 +1,7 @@
 ## unreleased
 * Add `RefundAuthHardDeclined` and `RefundAuthSoftDeclined` to validation errors
 * Add GraphQL ID to `CreditCardVerification`, `Customer`, `Dispute`, and `Transaction`
+* Add level 2 processing options `purchaseOrderNumber`, `taxAmount`, and `taxExempt` on transaction submitForSettlement
 
 ## 2.21.0
 * Add `AmountNotSupportedByProcessor` validation error to Transaction